add order: fg

+++ cloud-demo-pay-provider/src/main/java/com/sinbad/cloud/demo/pay/provider/mapper/PayMapper.java
@@ -10,7 +10,7 @@ import com.sinbad.cloud.demo.pay.provider.entry.po.PayPo;
 @Mapper
 public interface PayMapper {
 
-	int insertPayData(PayPo payPo);
+	long insertPayData(PayPo payPo);
 
 	PayPo getPayDataById(long payId);
 }
